Move ReferrerPolicy from page/SecurityPolicy.h to platform/ReferrerPolicy.h
